moves some modules
