+++ src/js/utils/backdrop/backdrop.js
@@ -1,6 +1,6 @@
 import './_backdrop.scss';
 import ScrollToggler from './toggler';
-import { getZindex, isInt } from '../utils';
+import { getZindex, isInt } from '../../utils';
 
 const scroll = new ScrollToggler();
